fix(tests): get the tests to run be reconfiguring build steps
