+++ .env.example
@@ -3,8 +3,10 @@ DATABASE_URL="postgresql://postgres:postgres@localhost:5432/postgres"
 PG_CLIENT_SSL="false"
 
 SESSION_SECRET="super-duper-s3cret"
+
 JWT_SECRET="OH GOD THIS IS SO INSECURE PLS CHANGE ME" # should be at least 32 characters
 JWT_ALGORITHM="HS256"
+JWT_ISSUER="https://api.my-opensensemap.com" # usually the base url of the api
 JWT_VALIDITY_MS=3600000 # 1 hour
 REFRESH_TOKEN_SECRET="I ALSO WANT TO BE CHANGED"
 REFRESH_TOKEN_ALGORTIHM="sha256"
