+++ app/routes/login.tsx
@@ -4,7 +4,7 @@ import ErrorMessage from "~/components/error-message";
 import { NavBar } from "~/components/nav-bar";
 
 import { verifyLogin } from "~/models/user.server";
-import { safeRedirect, validateEmail } from "~/utils";
+import { safeRedirect, validateEmail } from "~/utils.server";
 import { createUserSession, getUserId } from "~/utils/session.server";
 
 export async function loader({ request }: LoaderFunctionArgs) {
