+++ package.json
@@ -1,38 +1,24 @@
 {
-  "name": "frontend-d00c",
+  "name": "opensensemap",
   "type": "module",
   "private": true,
   "sideEffects": false,
   "scripts": {
+    "dev": "react-router dev",
     "build": "run-s build:*",
-    "build:drizzle": "npm run drizzle:generate",
-    "build:remix": "react-router build",
-    "build:server": "tsx ./other/build-server.ts",
-    "dev:build": "cross-env NODE_ENV=development npm run build:server -- --watch",
-    "dev:css": "cross-env NODE_ENV=development npm run generate:css -- --watch",
-    "dev": "cross-env NODE_ENV=development node ./server/dev-server.js",
-    "docker": "docker compose up -d",
-    "format": "prettier --write .",
-    "drizzle:generate": "drizzle-kit generate",
-    "drizzle:generate-custom": "drizzle-kit generate --custom",
-    "drizzle:migrate": "tsx ./db/migrate.ts",
-    "drizzle:drop": "drizzle-kit drop",
-    "drizzle:seed": "tsx ./db/seed.ts",
-    "drizzle:studio": "drizzle-kit studio",
-    "lint": "eslint .",
-    "setup": "npm run drizzle:generate && npm run drizzle:migrate && npm run drizzle:seed",
-    "start": "cross-env NODE_ENV=production node .",
-    "start:mocks": "cross-env NODE_ENV=production MOCKS=true node .",
+    "build:drizzle": "npm run db:generate",
+    "build:react-router": "react-router build",
+    "start": "cross-env NODE_ENV=production react-router-serve ./build/server/index.js",
     "test": "vitest",
-    "coverage": "vitest run --coverage",
-    "test:e2e": "npm run test:e2e:dev --silent",
-    "test:e2e:dev": "playwright test --ui",
-    "pretest:e2e:run": "npm run build",
-    "test:e2e:run": "cross-env CI=true playwright test",
-    "test:e2e:install": "npx playwright install --with-deps chromium",
-    "typecheck": "react-router typegen && tsc",
-    "validate": "run-p \"test -- --run\" lint typecheck test:e2e:run",
-    "components": "npx shadcn-ui"
+    "test:ui": "vitest --ui",
+    "test:coverage": "vitest run --coverage",
+    "db:push": "drizzle-kit push",
+    "db:generate": "drizzle-kit generate",
+    "db:migrate": "drizzle-kit migrate",
+    "components": "npx shadcn-ui",
+    "format": "prettier --write .",
+    "lint": "eslint .",
+    "typecheck": "react-router typegen && tsc"
   },
   "prettier": "@epic-web/config/prettier",
   "dependencies": {
@@ -162,6 +148,7 @@
     "@playwright/test": "^1.48.2",
     "@react-router/dev": "^7.0.0",
     "@react-router/fs-routes": "^7.1.1",
+    "@remix-run/dev": "^2.16.7",
     "@remix-run/eslint-config": "^2.15.2",
     "@tailwindcss/forms": "^0.5.3",
     "@tailwindcss/typography": "^0.5.13",
@@ -190,6 +177,7 @@
     "@types/supercluster": "^7.1.3",
     "@vitejs/plugin-react": "^4.0.4",
     "@vitest/coverage-v8": "^3.0.5",
+    "@vitest/ui": "^3.1.4",
     "autoprefixer": "^10.4.15",
     "cookie": "^1.0.2",
     "cssnano": "^7.0.6",
@@ -200,6 +188,7 @@
     "eslint-plugin-unicorn": "^56.0.1",
     "fs-extra": "^11.2.0",
     "happy-dom": "^15.7.4",
+    "jsdom": "^26.1.0",
     "msw": "^2.7.0",
     "npm-run-all": "^4.1.5",
     "postcss": "^8.5.2",
