--- tests/e2e/landing.spec.ts
@@ -1,6 +1,0 @@
-import { test } from "@playwright/test";
-
-// An empty test block to avoid errors for missing tests
-test("empty test block", async () => {
-  // No assertions or actions here
-});
